Extract settings data helpers
